+++ magnet/example_random_instance_predictor.py
@@ -10,11 +10,12 @@ class ExampleRandomInstancePredictor(InstancePredictor):
     Class to demonstrate a random per-instance stat prediction algorithm
 
     Example:
+        >>> from magnet.example_random_instance_predictor import *  # NOQA
         >>> import magnet
         >>> outputs = magnet.HelmOutputs.demo()
-        >>> suite_path = outputs.suites()[0].path
+        >>> helm_runs = outputs.suites()[0].runs()
         >>> predictor_instance = ExampleRandomInstancePredictor(num_eval_samples=5)
-        >>> predictor_instance(suite_path)
+        >>> predictor_instance(helm_runs=helm_runs)
     """
     def predict(self,
                 train_split: TrainSplit,
@@ -44,18 +45,19 @@ class ExampleRandomInstancePredictor(InstancePredictor):
 
         return predictions
 
+
 def main():
     parser = argparse.ArgumentParser(
         description="Run example random per-instance predictor")
 
-    parser.add_argument('helm_suite_path',
+    parser.add_argument('helm_runs',
                         type=str,
-                        help="Path to HELM run outputs for a suite (usually 'something/something/benchmark_output/runs/suite_name')")
+                        help="Pattern matching the set of HELM runs to evaluate on")
 
     args = parser.parse_args()
 
     predictor_instance = ExampleRandomInstancePredictor()
-    predictor_instance(args.helm_suite_path)
+    predictor_instance(helm_runs=args.helm_runs)
 
 
 if __name__ == "__main__":